[dlflib-js][encoder] Modified comments, moved float implementation to construct_floats.ts.
